#220 : minor refactoring
