+++ MysteryMud.Application.Commands/RegistryDependentCommands/OrderCommand.cs
@@ -13,7 +13,6 @@ using MysteryMud.Domain.Extensions;
 using MysteryMud.Domain.Services;
 using MysteryMud.GameData.Definitions;
 using MysteryMud.GameData.Enums;
-using System.Runtime.CompilerServices;
 
 namespace MysteryMud.Application.Commands.RegistryDependentCommands;
