+++ MysteryMud.Domain/Systems/AggressionSystem.cs
@@ -1,5 +1,7 @@
-﻿using MysteryMud.Core;
+﻿using Microsoft.Extensions.Logging;
+using MysteryMud.Core;
 using MysteryMud.Core.Bus;
+using MysteryMud.Domain.Helpers;
 using MysteryMud.Domain.Services;
 using MysteryMud.GameData.Events;
 
@@ -7,11 +9,13 @@ namespace MysteryMud.Domain.Systems;
 
 public class AggressionSystem
 {
+    private readonly ILogger _logger;
     private readonly ICombatService _combatService;
     private readonly IEventBuffer<AggressedEvent> _aggressed;
 
-    public AggressionSystem(ICombatService combatService, IEventBuffer<AggressedEvent> aggressed)
+    public AggressionSystem(ILogger logger, ICombatService combatService, IEventBuffer<AggressedEvent> aggressed)
     {
+        _logger = logger;
         _combatService = combatService;
         _aggressed = aggressed;
     }
@@ -21,7 +25,10 @@ public class AggressionSystem
     {
         foreach (ref var evt in _aggressed.GetAll())
         {
-            _combatService.EnterCombat(state, evt.Source, evt.Target);
+            if (CharacterHelpers.SameRoom(evt.Source, evt.Target))
+                _combatService.EnterCombat(state, evt.Source, evt.Target);
+            else
+                _logger.LogError("AggressionSystem: Source {Source} and Target {Target} are not in the same room, cannot enter combat", evt.Source, evt.Target);
         }
     }
 }
