+++ MysteryMud.Domain/Systems/TimedEffectSystem.cs
@@ -12,7 +12,6 @@ using MysteryMud.Domain.Helpers;
 using MysteryMud.Domain.Services;
 using MysteryMud.GameData.Enums;
 using MysteryMud.GameData.Events;
-using MysteryMud.GameData.Time;
 
 namespace MysteryMud.Domain.Systems;
